Fix/refactor NodeScoreIterator, BoundedLongHeap, and GrowableLongHeap bulk addition implementations (#433)
* Fix boundary check on BoundedLongHeap::pushAll

* Breaking change: refactor NodeScoreIterator

The methods in the NodeScoreIterator
had opposite semantics from the NodeQueue.
One key problem was that nextScore
advanced the iterator. I propose instead
that we align with the NodeQueue names
and add topScore along with pop as the
key methods and pop will advance the
iterator. This is a breaking change
both in terms of the API and the semantics,
but because this interface was only
just released last week and we are
also breaking the compile-compatibility,
I view this as pretty low risk.

* Fix missed requirement that heap is 1-based, not 0-based

* Rename pushAll to pushN to match actual semantics

* Fix test assertion

* Rename pushN to pushMany; update javadocs